+++ src/contact/keys.ts
@@ -19,5 +19,6 @@ export const contactTriggers = {
   updatedContactTrigger: 'contact_update_contact_trigger',
   contactListDropdown: 'contact_list_contacts',
   personListDropdown: 'contact_list_persons',
-  companyListDropdown: 'contact_list_companies'
+  companyListDropdown: 'contact_list_companies',
+  contactFieldsDropdown: 'contact_fields_dropdown'
 }