+++ src/deal/updatedDeal.trigger.ts
@@ -13,10 +13,10 @@ const listDealsByUpdatedAt = async (z: ZObject, bundle: Bundle) => {
     'updated_at',
     []
   )(z, bundle)
-  return findAndRemapOnlyUpdatedItems(deals)
+  return findAndRemapOnlyUpdatedItems(deals, bundle.inputData.trigger_field)
 }
 
-const DealUpdatedTrigger: ZapierItem = {
+const UpdatedDealTrigger: ZapierItem = {
   key: dealTriggers.updatedDealTrigger,
   noun: 'Deal',
   display: {
@@ -26,6 +26,16 @@ const DealUpdatedTrigger: ZapierItem = {
   operation: {
     // Resource cannot be used here, because of different output fields (deduplication)
     sample: dealSample,
+    inputFields: [
+      {
+        key: 'trigger_field',
+        label: 'Field to monitor the updates',
+        helpText: 'Trigger will work only when selected field gets updated. Leave empty to trigger on any change.',
+        required: false,
+        type: 'string',
+        dynamic: `${dealTriggers.dealFieldsDropdown}.id.name`
+      }
+    ],
     outputFields: [
       ...deduplicationOutputFields,
       ...dealCommonOutputFields
@@ -34,4 +44,4 @@ const DealUpdatedTrigger: ZapierItem = {
   }
 }
 
-export default DealUpdatedTrigger
+export default UpdatedDealTrigger