+++ src/utils/__test__/deduplication.test.ts
@@ -1,4 +1,5 @@
 import {findAndRemapOnlyStageUpdatedItems, findAndRemapOnlyUpdatedItems, remapDeduplication} from '../deduplication'
+import {assertDeduplicationIds} from '../testHelpers'
 
 describe('findAndRemapOnlyUpdatedItems', () => {
   it('should return empty array if input is empty', () => {
@@ -35,7 +36,7 @@ describe('findAndRemapOnlyStageUpdatedItems', () => {
 })
 
 describe('remapDeduplication', () => {
-  it('should properly remap ids', () => {
+  it('should properly remap ids for any field', () => {
     const deals = [
       {id: 1, name: 'Uzi deal', field_10: 'one'},
       {id: 2, name: 'Bart deal', field_10: 'two'},
@@ -45,5 +46,101 @@ describe('remapDeduplication', () => {
     expect(deduplicatedItems).toHaveLength(2)
     expect(deduplicatedItems.map(item => item.id)).toEqual(['1_one', '2_two'])
     expect(deduplicatedItems.map(item => item.entity_original_id)).toEqual([1, 2])
+  })
+
+  it('should use nulls when field is not present', () => {
+    const deals = [
+      {id: 100, field_10: 'Value'},
+      {id: 200},
+      {id: 300, field_10: undefined},
+    ]
+
+    const deduplicatedItems = remapDeduplication(deals, 'field_10')
+    expect(deduplicatedItems).toHaveLength(3)
+    assertDeduplicationIds(
+      deduplicatedItems,
+      [100, 200, 300],
+      ['100_Value', '200_null', '300_null'],
+    )
+  })
+
+  it('should use nulls when fields is nulled', () => {
+    const deals = [
+      {id: 100, field_10: 'Value'},
+      {id: 200, field_10: null},
+      {id: 300, field_10: null},
+    ]
+
+    const deduplicatedItems = remapDeduplication(deals, 'field_10')
+    expect(deduplicatedItems).toHaveLength(3)
+    assertDeduplicationIds(
+      deduplicatedItems,
+      [100, 200, 300],
+      ['100_Value', '200_null', '300_null'],
+    )
+  })
+
+  it('should properly use nested fields for deduplication id', () => {
+    const deals = [
+      {id: 100, custom_fields: {field: 'a'}},
+      {id: 200, custom_fields: {}},
+      {id: 300},
+    ]
+
+    const deduplicatedItems = remapDeduplication(deals, 'custom_fields.field')
+    expect(deduplicatedItems).toHaveLength(3)
+    assertDeduplicationIds(
+      deduplicatedItems,
+      [100, 200, 300],
+      ['100_a', '200_null', '300_null'],
+    )
+  })
+
+  it('should properly use nested fields with dots', () => {
+    const deals = [
+      {id: 100, custom_fields: {'field.with.dot': 'XXX'}},
+      {id: 200, custom_fields: {}},
+      {id: 300},
+    ]
+
+    const deduplicatedItems = remapDeduplication(deals, 'custom_fields.field.with.dot')
+    expect(deduplicatedItems).toHaveLength(3)
+    assertDeduplicationIds(
+      deduplicatedItems,
+      [100, 200, 300],
+      ['100_XXX', '200_null', '300_null'],
+    )
+  })
+
+  it('should properly use arrays as keys', () => {
+    const deals = [
+      {id: 100, tags: ['a', 'b', 'c']},
+      {id: 200, tags: [1, 2, 3]},
+      {id: 300},
+    ]
+
+    const deduplicatedItems = remapDeduplication(deals, 'tags')
+    expect(deduplicatedItems).toHaveLength(3)
+      assertDeduplicationIds(
+      deduplicatedItems,
+      [100, 200, 300],
+      ['100_a,b,c', '200_1,2,3', '300_null'],
+    )
+  })
+
+  it('should properly use object as keys', () => {
+    const deals = [
+      {id: 100, address: {street: 'a'}},
+      {id: 200, address: {}},
+      {id: 300},
+    ]
+
+    const deduplicatedItems = remapDeduplication(deals, 'address')
+    expect(deduplicatedItems).toHaveLength(3)
+    assertDeduplicationIds(
+      deduplicatedItems,
+      [100, 200, 300],
+      ['100_{"street":"a"}', '200_{}', '300_null'],
+    )
   })
 })